+++ src/ape/api/networks.py
@@ -12,7 +12,7 @@ from eth_utils import keccak, to_int
 from ethpm_types import ContractType, HexBytes
 from ethpm_types.abi import ABIType, ConstructorABI, EventABI, MethodABI
 
-from ape._compat import BaseModel
+from ape._pydantic_compat import BaseModel
 from ape.exceptions import (
     NetworkError,
     NetworkMismatchError,
