+++ src/ape/api/query.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, Iterator, List, Optional, Sequence, Set, Type, Uni
 
 from ethpm_types.abi import EventABI, MethodABI
 
-from ape._compat import BaseModel, NonNegativeInt, PositiveInt, root_validator
+from ape._pydantic_compat import BaseModel, NonNegativeInt, PositiveInt, root_validator
 from ape.api.transactions import ReceiptAPI, TransactionAPI
 from ape.logging import logger
 from ape.types import AddressType
