+++ src/ape/managers/config.py
@@ -3,7 +3,7 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Dict, Generator, List, Optional, Union
 
-from ape._compat import root_validator
+from ape._pydantic_compat import root_validator
 from ape.api import ConfigDict, DependencyAPI, PluginConfig
 from ape.exceptions import ConfigError, NetworkError
 from ape.logging import logger
