+++ src/ape_ethereum/ecosystem.py
@@ -17,7 +17,7 @@ from eth_utils import (
 from ethpm_types import ContractType, HexBytes
 from ethpm_types.abi import ABIType, ConstructorABI, EventABI, MethodABI
 
-from ape._compat import Field, validator
+from ape._pydantic_compat import Field, validator
 from ape.api import BlockAPI, EcosystemAPI, PluginConfig, ReceiptAPI, TransactionAPI
 from ape.api.networks import LOCAL_NETWORK_NAME
 from ape.contracts.base import ContractCall
