+++ src/ape_plugins/utils.py
@@ -3,7 +3,7 @@ import sys
 from typing import List, Optional, Tuple
 
 from ape.__modules__ import __modules__
-from ape._compat import root_validator
+from ape._pydantic_compat import root_validator
 from ape.logging import ApeLogger
 from ape.plugins import clean_plugin_name
 from ape.utils import BaseInterfaceModel, cached_property, get_package_version, github_client
